feat: 투표시 좋아요 구현

+++ src/main/java/link/sendwish/backend/dtos/chat/ChatVoteResponseDto.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ChatVoteResponseDto {
     private Long itemId;
-    private Long like;
+    private Integer like;
 }
